Develop

+++ ATS.Repository/DAO/UserRepository.cs
@@ -31,8 +31,8 @@ namespace ATS.Repository.DAO
                             userInfo.LName = input.LName;
                             userInfo.Mobile = input.Mobile;
                             userInfo.StatusId = input.StatusId;
-                            userInfo.UserId = Guid.NewGuid();
-                            userInfo.UserTypeId = input.UserTypeId;
+                            userInfo.UserId = Guid.NewGuid();      
+                            userInfo.RoleTypeId = input.RoleTypeId;
 
                             context.UserInfo.Add(userInfo);
 
@@ -42,6 +42,7 @@ namespace ATS.Repository.DAO
                             if (isCreated)
                             {
                                 UserCredential userCredential = new UserCredential();
+                                userCredential.Id = Guid.NewGuid();
                                 userCredential.EmailId = input.Email;
                                 userCredential.CurrPassword = input.UserCredentials[0].CurrPassword;
                                 userCredential.UserId = userInfo.UserId;
@@ -80,15 +81,31 @@ namespace ATS.Repository.DAO
                 {
                     try
                     {
-                        UserInfo userInfo = context.UserInfo.AsNoTracking().Where(x => x.UserId == input.UserId).FirstOrDefault();
-                        if (userInfo != null)
+                        UserCredential userCredential = context.UserCredential.AsNoTracking().Where(x => x.UserId == input.UserId).FirstOrDefault();
+                        if (userCredential != null)
                         {
-                            context.UserInfo.Remove(userInfo);
+                            context.UserCredential.Remove(userCredential);
                             context.SaveChanges();
-
-                            dbContextTransaction.Commit();
                             isDeleted = true;
-                        }
+
+                            UserInfo userInfo = context.UserInfo.AsNoTracking().Where(x => x.UserId == input.UserId).FirstOrDefault();
+                            if (userInfo != null && isDeleted == true)
+                            {
+                                context.UserInfo.Remove(userInfo);
+                                context.SaveChanges();
+                                isDeleted = true;
+                            }
+                            else
+                            {
+                                isDeleted = false;
+                            }
+
+                        }                        
+                        if (isDeleted)
+                        {
+                            dbContextTransaction.Commit();
+                        }                        
+
                     }
                     catch
                     {
@@ -115,7 +132,7 @@ namespace ATS.Repository.DAO
                 }
                 return userInfo;
             }
-        }
+        } 
 
         public List<UserInfoModel> Select(Func<UserInfoModel, bool> condition)
         {
@@ -124,7 +141,7 @@ namespace ATS.Repository.DAO
             {
                 try
                 {
-                    Select(context, condition);
+                    userInfos=Select(context, condition);
                 }
                 catch
                 {
@@ -140,6 +157,8 @@ namespace ATS.Repository.DAO
 
             var query = (from x in context.UserInfo.AsNoTracking()
                          join y in context.TypeDef on x.RoleTypeId equals y.TypeId
+                         join z in context.UserCredential on x.UserId equals z.UserId
+
                          select new UserInfoModel
                          {
                              UserId = x.UserId,
@@ -152,10 +171,11 @@ namespace ATS.Repository.DAO
                              LName = x.LName,
                              Mobile = x.Mobile,
                              StatusId = x.StatusId,
-                             UserTypeId = x.UserTypeId,
+                             //UserTypeId = x.UserTypeId,
                              RoleTypeId = x.RoleTypeId,
                              RoleDescription = y.Description,
-                             RoleValue = y.Value
+                             RoleValue = y.Value,
+                             CurrPassword=z.CurrPassword                             
                          });
             if (condition != null)
             {
@@ -185,7 +205,7 @@ namespace ATS.Repository.DAO
                         {
                             userInfo.LastUpdatedBy = input.LastUpdatedBy;
                             userInfo.LastUpdatedDate = input.LastUpdatedDate;
-                            userInfo.UserTypeId = input.UserTypeId;
+                            //userInfo.UserTypeId = input.UserTypeId;
                             userInfo.FName = input.FName;
                             userInfo.LName = input.LName;
                             userInfo.Mobile = input.Mobile;
